requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
boto3==1.26.119
botocore==1.29.119
branca==0.6.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.1
django-bootstrap4==22.3
django-custom-user==1.1
django-filter==22.1
django-heroku==0.3.1
django-ipware==5.0.0
django-pandas==0.6.6
django-s3-storage==0.13.11
django-storages==1.13.2
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
folium==0.13.0
fonttools==4.39.3
future==0.18.2
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
numpy==1.23.5
packaging==23.1
pandas==2.0.1
Pillow==9.2.0
plotly==5.14.1
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
ratelim==0.1.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.12
whitenoise==6.2.0
